+++ x/stablecoin/keeper/collateral_ratio_test.go
@@ -129,7 +129,7 @@ func TestSetCollRatioUpdate(t *testing.T) {
 			name:              "Collateral price is higher than stable",
 			inCollRatio:       sdk.MustNewDecFromStr("0.8"),
 			price:             sdk.MustNewDecFromStr("1.1"),
-			expectedCollRatio: sdk.MustNewDecFromStr("0.7975"),
+			expectedCollRatio: sdk.MustNewDecFromStr("0.8025"),
 			expectedPass:      true,
 		},
 		{
@@ -150,7 +150,7 @@ func TestSetCollRatioUpdate(t *testing.T) {
 			name:              "Collateral price is lower than stable",
 			inCollRatio:       sdk.MustNewDecFromStr("0.8"),
 			price:             sdk.MustNewDecFromStr("0.9"),
-			expectedCollRatio: sdk.MustNewDecFromStr("0.8025"),
+			expectedCollRatio: sdk.MustNewDecFromStr("0.7975"),
 			expectedPass:      true,
 		},
 	}